Make the legacyPackages schema lazier
Previously we tried to filter out attributes that are neither a
derivation or a `recurseForDerivations` attrset. However, this meant
that we had to evaluate every attribute just to get the list of
attributes. Thus, since
https://github.com/NixOS/nix/pull/8892/commits/5fddc13dac8a0f6dfe0573feb538526e432dc27e
(where Nix started using flake schemas for regular evaluations), an
operation like `nix build nixpkgs#hello` would be much slower than
needed and print a lot of warnings.

The solution is to not filter out attributes. Instead, we just return
a value `{ what = "unknown"; }` for unrecognized attributes.